chore: Move to org.eclipse package
