--- vss-processor/src/main/kotlin/com/etas/kuksa/vssprocessor/VssDefinition.kt
@@ -1,5 +1,0 @@
-package com.etas.kuksa.vssprocessor
-
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.SOURCE)
-annotation class VssDefinition(val vssDefinitionPath: String)
