--- vss-processor/src/main/kotlin/com/etas/kuksa/vssprocessor/VssSpecification.kt
@@ -1,12 +1,0 @@
-package com.etas.kuksa.vssprocessor
-
-interface VssSpecification {
-    val uuid: String
-    val vssPath: String
-    val description: String
-    val type: String
-}
-
-interface VSSProperty<T> {
-    val value: T
-}
