+++ vss-processor/src/main/kotlin/org/eclipse/kuksa/vssprocessor/VssSpecification.kt
@@ -17,19 +17,15 @@
  *
  */
 
-package org.eclipse.kuksa
+package org.eclipse.kuksa.vssprocessor
 
-import org.eclipse.kuksa.proto.v1.Types.DataEntry
+interface VssSpecification {
+    val uuid: String
+    val vssPath: String
+    val description: String
+    val type: String
+}
 
-/**
- * The Observer used to notify about changes to subscribed properties.
- */
-fun interface PropertyObserver {
-    /**
-     * Will be triggered when the underlying vssPath changed it's value.
-     *
-     * @param vssPath the underlying vssPath
-     * @param updatedValue the updated value
-     */
-    fun onPropertyChanged(vssPath: String, updatedValue: DataEntry)
+interface VSSProperty<T> {
+    val value: T
 }
