+++ vss-processor/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,1 +1,1 @@
-com.etas.kuksa.vssprocessor.VssDefinitionProcessorProvider
+org.eclipse.kuksa.vssprocessor.VssDefinitionProcessorProvider
